pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pcap-maker"
version = "0.1.0"
description = "A tool to create pcap files."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Colm O'Shea", email = "colmo23@gmail.com" },
]
dependencies = [
    "dpkt==1.9.8",
    "Flask==3.0.0",
    "Werkzeug==3.0.0",
]

[project.urls]
Homepage = "https://github.com/colmo23/pcap-maker" 

[project.optional-dependencies]
dev = [
    "pytest",
    "pylint",
    "flake8",
    "black",
    "autopep8",
    "tox",
    "pytest-playwright",
]

[tool.pytest.ini_options]
addopts = "--ignore=dpkt"
python_files = "*.py"
python_functions = "test"
norecursedirs = ".tox .git *.egg-info __pycache__ dist build dpkt"
testpaths = "."

[tool.flake8]
max-line-length = 1000
exclude = ".git,__pycache__,docs/source/conf.py,old,build,dist,dpkt,bottle.py"

[tool.black]
line-length = 1000
